fix(kv): launch BF16 prefix append with BF16 values

+++ src/ops/kv_cache/append/launch.cu
@@ -132,7 +132,7 @@ void launch_paged(const Tensor& k, const Tensor& v, const Tensor& positions, con
     validate_plan(k, plan);
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k_pages.data);
-    auto* cache_v       = static_cast<__half*>(cache.v_pages.data);
+    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v_pages.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
@@ -153,7 +153,7 @@ void launch_cyclic_profile(const Tensor& k, const Tensor& v, const Tensor& posit
                            const KVCacheAppendPrefixPlan& plan, cudaStream_t stream) {
     if (plan.max_count == 0) return;
     auto* cache_k       = static_cast<__nv_bfloat16*>(cache.k.data);
-    auto* cache_v       = static_cast<__half*>(cache.v.data);
+    auto* cache_v       = static_cast<__nv_bfloat16*>(cache.v.data);
     const auto* input_k = static_cast<const __nv_bfloat16*>(k.data);
     const auto* input_v = static_cast<const __nv_bfloat16*>(v.data);
     const auto* pos     = static_cast<const std::int32_t*>(positions.data);
